Reformat create.tsx
